requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
pydantic>=2.6
psutil>=5.9
httpx>=0.26
pytest>=8.0

PyYAML>=6.0
